Fix C++ interface build by extracting texture callback types
Move TextureLoadData and TextureLoadCallback from TextureManager.hpp
to a separate TextureTypes.hpp header. This allows ProjectM.hpp to
include only the necessary types without pulling in internal
dependencies that aren't installed.
