+++ src/libprojectM/CMakeLists.txt
@@ -187,6 +187,7 @@ if(ENABLE_INSTALL)
 
         install(FILES
                 Renderer/RenderContext.hpp
+                Renderer/TextureTypes.hpp
                 DESTINATION "${PROJECTM_INCLUDE_DIR}/projectM-4/Renderer"
                 COMPONENT Devel
                 )
